flash 回滚配置文件丢失

+++ src/main/resources/templates/updateScript/sh/flashRollbackSh.ftl
@@ -17,6 +17,9 @@ else
         if [ "$fileNameNoSuffix" = "resource" ]; then
             continue;
         fi
+        if [ "$fileNameNoSuffix" = "config" ]; then
+            continue;
+        fi
         rm -r $file
     done
